==> data_process/00_data_cleaning.ipynb <==

import pandas as pd
from numpy.random import random
from chatbot import QWENChatBot
from pandarallel import pandarallel
import time
import random
from concurrent.futures import ThreadPoolExecutor
from typing import List

data = pd.read_excel('raw_data/LV_20250101-0206.xlsx')
data.head()

#content null
data = data[data['内容'].notnull()]

system_message = """
你是一名数据标注专家，现在需要对用户发布的关于 LVMH 产品及品牌的小红书帖子进行分类。你的任务是根据以下规则，将帖子分类为四种类型之一：

规则：
1. 如果帖子是营销贴/广告贴，请标注为 **1**。
   - 特征：以推广产品或鼓励购买为目的，包含促销、价格、优惠信息、购买链接或显而易见的商业宣传语言。
   - 特征：使用品牌合作或官方话术，明确商业性质内容。
2. 如果帖子与明星相关性较高，请标注为 **2**。
   - 特征：明星代言分析、明星街拍讨论、明星推广行为等
   - 特征：明确提到某明星（如“张某某代言”、明星机场图）
3. 如果帖子是普通用户对 LVMH 相关产品的真实反馈，请标注为 **0**。
   - 特征：分享个人真实体验或意见，通常语气更加自然，没有强烈的推销目的。
   - 特征：通常描述个人对产品的使用感受、优缺点评价。
4. 如果无法分类，请标注为 **-1**。
要求：
输出结果必须是整数，范围严格为 -1、0、1 或 2，不允许出现范围以外的任何内容，不要添加解释。
"""
chatbot = QWENChatBot(system_message=system_message)

'''
# 定义一个函数用于将帖子分类
def classify_post(content: str) -> dict:
    # 调用 QWEN 模型，得到响应
    try:
        response = chatbot.ask(content)
    except Exception as e:
        response = f"Error: {str(e)}"
    
    return response
'''
def classify_post_thread_safe(content: str) -> int:
    try:
        return chatbot.ask(content)
    except Exception as e:
        return f"Error: {e}"
def classify_posts_concurrently(posts: List[str], max_workers: int = 5) -> List[int]:
    results = []
    with ThreadPoolExecutor(max_workers=max_workers) as executor:
        # 将所有帖子分配给线程池并行处理
        futures = list(executor.map(classify_post_thread_safe, posts))
        results.extend(futures)
    return results

# 使用线程池处理 DataFrame 数据
data['classification'] = classify_posts_concurrently(data['内容'].tolist(), max_workers=16)


'''
# 批量分类：为每一行调用 classify 方法
data['classification'] = data['内容'].apply(classify_post)
'''
data

data.classification.value_counts()

filtered_data = data[data.classification == '0']
filtered_data

filtered_data.to_excel('processed_data/raw_data_filtered.xlsx',index = False)
