==> data_process/02_data_tagging_processing.ipynb <==
import os
import sys
project_path = os.path.abspath(os.path.join(os.getcwd(), '..'))  # Go up one level
sys.path.append(project_path)
import pandas as pd
from numpy.random import random
from utils.chatbot import QWENChatBot
import time
import random
import json
import numpy as np
from concurrent.futures import ThreadPoolExecutor

data_tagged = pd.read_excel('../processed_data/01_tagged_raw.xlsx')
data_tagged


def safe_json_parser(json_string):
    try:
        # Try parsing the JSON
        return json.loads(json_string)
    except json.JSONDecodeError:
        # If an error occurs, skip/return None
        return None

# Apply the function to the column of JSON data
data_tagged['parsed_json'] = data_tagged['raw_tagging'].apply(safe_json_parser)
# Optionally, drop rows where parsing failed (parsed_json is None)
data_tagged = data_tagged.dropna(subset=['parsed_json']).reset_index(drop=True)

#data_tagged['raw_tagging'] = data['raw_tagging'].apply(json.loads)
data_tagged['Age_Range'] = data_tagged['parsed_json'].apply(lambda x: x.get('Age_Range', np.nan) if x.get('Age_Range', '') != '' else np.nan)
data_tagged['Gender'] = data_tagged['parsed_json'].apply(lambda x: x.get('Gender', np.nan) if x.get('Gender', '') != '' else np.nan)
data_tagged['City_Tier'] = data_tagged['parsed_json'].apply(lambda x: x.get('City_Tier', np.nan) if x.get('City_Tier', '') != '' else np.nan)
data_tagged['Purchase_Purpose'] = data_tagged['parsed_json'].apply(lambda x: x.get('Purchase_Purpose', np.nan) if x.get('Purchase_Purpose', '') != '' else np.nan)
data_tagged['Purchase_Situation'] = data_tagged['parsed_json'].apply(lambda x: x.get('Purchase_Situation', np.nan) if x.get('Purchase_Situation', '') != '' else np.nan)
data_tagged['Product_Mention'] = data_tagged['parsed_json'].apply(lambda x: x.get('Product_Mention', np.nan) if x.get('Product_Mention', []) != [] else np.nan)
data_tagged['Competitor'] = data_tagged['parsed_json'].apply(lambda x: x.get('Competitor', np.nan) if x.get('Competitor', []) != [] else np.nan)
data_tagged['Keywords_Design_Elements'] = data_tagged['parsed_json'].apply(lambda x: x.get('Keywords', {}).get('Design_Elements', np.nan) if x.get('Keywords', {}).get('Design_Elements', []) != [] else np.nan)
data_tagged['Keywords_Design_Style'] = data_tagged['parsed_json'].apply(lambda x: x.get('Keywords', {}).get('Design_Style', np.nan) if x.get('Keywords', {}).get('Design_Style', []) != [] else np.nan)
data_tagged['Keywords_Functionality'] = data_tagged['parsed_json'].apply(lambda x: x.get('Keywords', {}).get('Functionality', np.nan) if x.get('Keywords', {}).get('Functionality', []) != [] else np.nan)
data_tagged['Keywords_Other'] = data_tagged['parsed_json'].apply(lambda x: x.get('Keywords', {}).get('Other', np.nan) if x.get('Keywords', {}).get('Other', []) != [] else np.nan)
data_tagged['Emotion'] = data_tagged['parsed_json'].apply(lambda x: x.get('Emotion', np.nan) if x.get('Emotion', '') != '' else np.nan)
data_tagged['Past_Purchase'] = data_tagged['parsed_json'].apply(lambda x: x.get('Past_Purchase', np.nan) if x.get('Past_Purchase', '') != '' else np.nan)
data_tagged['Tone'] = data_tagged['parsed_json'].apply(lambda x: x.get('Tone', np.nan) if x.get('Tone', '') != '' else np.nan)

data_tagged


data_tagged.isnull().sum()/len(data_tagged)


data_tagged.Purchase_Purpose.value_counts()

data_tagged.Purchase_Situation.value_counts()[:10]

data_tagged.columns

cols = ['Age_Range', 'Gender', 'City_Tier',
       'Purchase_Purpose', 'Purchase_Situation', 'Product_Mention',
       'Competitor', 'Keywords_Design_Elements', 'Keywords_Design_Style',
       'Keywords_Functionality', 'Emotion', 'Past_Purchase',
       'Tone']

data_tagged['Product_Mention'].value_counts()

length = 0
for col in cols:
    length+= len(data_tagged[col].value_counts())
length

data_tagged.to_excel('../processed_data/02_tagged_processed.xlsx',index=False)

from collections import Counter
def count_word_freq(column):
    # Flatten all lists from the column into a single list, ignoring NaN values
    all_keywords = [word for keywords in column.dropna() for word in keywords]
    
    # Count the frequency of each word using Counter
    return Counter(all_keywords)
# Function to get top N frequent keywords
def get_top_n(counter_obj, n=5):
    return counter_obj.most_common(n)


# Count word frequencies for each subcategory
freq_product = count_word_freq(data_tagged['Product_Mention'])
freq_competitor = count_word_freq(data_tagged['Competitor'])
freq_design_elements = count_word_freq(data_tagged['Keywords_Design_Elements'])
freq_design_style = count_word_freq(data_tagged['Keywords_Design_Style'])
freq_functionality = count_word_freq(data_tagged['Keywords_Functionality'])
freq_other = count_word_freq(data_tagged['Keywords_Other'])

# Display top 5 keywords for each subcategory
print("Design Elements Word Frequency (Top 5):")
print(get_top_n(freq_product, 10))
print("Design Elements Word Frequency (Top 5):")
print(get_top_n(freq_competitor, 10))
print("Design Elements Word Frequency (Top 5):")
print(get_top_n(freq_design_elements, 10))

print("\nDesign Style Word Frequency (Top 5):")
print(get_top_n(freq_design_style, 10))

print("\nFunctionality Word Frequency (Top 5):")
print(get_top_n(freq_functionality, 10))

print("\nOther Word Frequency (Top 5):")
print(get_top_n(freq_other, 15))
